LayoutTests/fast/canvas/getCSSCanvasContext-create-and-request-expected.txt

Document.getCSSCanvasContext() should return null if the context was created with a different type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Creating CSSCanvasContext with type = 2d
PASS context != null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
Creating CSSCanvasContext with type = bitmaprenderer
PASS context != null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
Creating CSSCanvasContext with type = webgl
PASS context != null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
Creating CSSCanvasContext with type = webgl2
PASS context != null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS type == type2 ? context2 == context : context2 == null is true
PASS successfullyParsed is true

TEST COMPLETE
